Add redirection to `search-result`.
Redirection was done from SearchOptions component. The states in SearchOptions component may be lifted up to SeachBar component.
