+++ web/src/components/SearchResults.jsx
@@ -1,5 +1,21 @@
 import React from 'react';
+import { useLocation } from 'react-router-dom';
 
-const SearchResults = () => <p>Search Results</p>;
+const SearchResults = () => {
+  const location = useLocation();
+  const queryParamsParser = new URLSearchParams(location.search);
+  const queryParams = {};
+  queryParamsParser.forEach((v, k) => {
+    queryParams[k] = v;
+  });
+
+  console.log(queryParams);
+
+  return (
+    <p>
+      Search Results
+    </p>
+  );
+};
 
 export default SearchResults;
